Added QR Code Generator
The function of the QR Code Generator is to generate a QR Code containing the data or information filled out and submitted to the program that can be viewed and copied by scanning it through a camera with a built-in QR Code Scanner or a QR Code Scanner app. Huge thanks to Raffael Herrmann for the NuGet Package called "QRCoder" for making it possible to generate QR Codes within the program.