Added QR Code Scanner
The added feature allows the user to scan QR Code/s through the webcam and the program automatically fill out the text box with the data obtained from the scanned QR Code/s. Huge thanks to the creators of AForge.Video. DirectShow and ZXing.Net NuGet Packages for bringing this functionality to the program.